dna-mutation-simulator/pyproject.toml
[project]
name = "replication-mutation-sim"
version = "1.0.0"
description = "DNA replication, mutation and repair simulation pipeline"
requires-python = ">=3.10"

dependencies = [
    "biopython",
    "python-dotenv",
    "matplotlib",
    "numpy",
    "pandas",
]